refactor(dashboard): fold catalog helpers into utils
Remove lib/catalogs.ts; keep the same localStorage and URI helpers on the existing utils module.
